move MeetingAvailability to its own file create styles directory with SelectAttendees and MemberCheckbox stylesheets move MemberCheckbox into its own file style close to figma
