+++ src/components/Calendar/MeetingModal/MeetingModal.tsx
@@ -235,7 +235,6 @@ export const MeetingModal = () => {
               />
 
               <SelectAttendees
-                authUser={authUser}
                 attendees={attendees}
                 dateFields={dateFields}
                 inviteAll={inviteAll}
