--- src/components/Calendar/MeetingModal/SelectAttendees.scss
@@ -1,1 +1,0 @@
-// TODO: Move selectAttendees scss to this file
