+++ src/components/Calendar/MeetingModal/styles/MemberCheckbox.scss
@@ -1,0 +1,12 @@
+.member-checkbox {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 70%;
+
+  p {
+    color: #616161;
+    font-style: italic;
+    font-weight: 400;
+  }
+}
