+++ src/components/Calendar/MeetingModal/styles/SelectAttendees.scss
@@ -1,0 +1,17 @@
+.select-attendees-section {
+  display: flex;
+  flex-direction: row;
+  grid-area: selectAttendees;
+  margin-bottom: 32px;
+  .select-attendees-wrapper {
+    width: 100%
+  }
+  :first-child {
+    margin-right: 20px;
+  }
+
+  .select-attendees-helper-text {
+    font-size: 12px;
+    font-style: italic;
+  }
+}
